requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain_core>=0.1.5
langchain-community>=0.0.10
langgraph>=0.0.20
tavily-python>=0.2.6
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.10.0
pytest>=7.0.0
pytest-mock>=3.10.0